HA-3 - Cleaned up weather layout, removed debug statements dev - Moved HomeServerDao to dao directory
